(feat) change previously added comments via inline buttons

+++ dreamprf_diary_bot/bot.py
@@ -1,5 +1,6 @@
 import json
 import logging
+import os.path
 import re
 from datetime import datetime, timedelta, timezone
 
@@ -40,7 +41,7 @@ async def send_welcome(message: types.Message):
 
 
 @dp.message_handler(commands='help')
-async def send_welcome(message: types.Message):
+async def send_help(message: types.Message):
     response = static.help_message
     await message.reply(response, reply=False)
 
@@ -73,6 +74,7 @@ async def post_night_wake(message: types.Message, state: FSMContext):
     write_range = get_night_wake_range(sheet, time)
     logging.info(f'Night wake-up at {time_str}. Log into range {write_range}')
     sheet.update(write_range, [[time_str]])
+    await state.update_data(last_night_time=time_str)
     await message.reply(f"Записано НП в {time_str}", reply=False)
 
 
@@ -84,20 +86,51 @@ async def cmd_night(message: types.Message, state: FSMContext):
     await state.finish()
 
 
+@dp.message_handler(commands='last', state=NightRegister.start)
+async def last_night(message: types.Message):
+    sid = get_spreadsheet_id(message.from_user.id)
+    sheet = Sheet(spreadsheet_id=sid)
+    dates = sheet.get_data(sheet_range=DATES_RANGE)
+    time = get_datetime_to_fill()
+    col_to_write = dates[0].index(time.strftime('%d.%m.%Y'))
+    response = sheet.get_data(f'R{NIGHT_ROW_INDEX + 1}C{col_to_write + 1}:R{NIGHT_ROW_INDEX+NIGHT_LENGTH+1}C{col_to_write+2}')
+    response = [' - '.join(line) for line in response]
+    # await message.reply(response, reply=False)
+
+    keyboard = types.InlineKeyboardMarkup()
+    for r in response:
+        time_str = r.split(' - ')[0]
+        keyboard.add(types.InlineKeyboardButton(text=r, callback_data=f'edit_comment_{time_str.replace(":", "_")}'))
+
+    await message.answer("Выберите просыпание для исправления комментария", reply_markup=keyboard)
+
+
+@dp.callback_query_handler(Text(startswith='edit_comment_'), state=NightRegister.start)
+async def edit_comment(call: types.CallbackQuery, state: FSMContext):
+    logging.info(f'Callback catched: {call.data=}')
+    time_str = call.data.split('_', 2)[-1].replace('_', ':')
+    await state.update_data(last_night_time=time_str)
+    await call.message.reply(f'Введите новый комментарий для просыпания в {time_str}', reply=False)
+    await call.answer()
+
+
 @dp.message_handler(state=NightRegister.start)
 async def note(message: types.Message, state: FSMContext):
     sid = get_spreadsheet_id(message.from_user.id)
     if not sid:
-        await message.reply('Не удалось найти ссылку на таблицу для вас. Отправьте ссылку на таблицу боту, не находясь в режиме записи пробуждений')
+        await message.reply('Не удалось найти ссылку на таблицу для вас. Отправьте ссылку на таблицу боту, '
+                            'не находясь в режиме записи пробуждений')
         return
     sheet = Sheet(spreadsheet_id=sid)
     cmt = message.text
+    stored_data = await state.get_data()
+    last_night_time = stored_data.get('last_night_time', None)
     time = get_datetime_to_fill()
-    write_range = get_night_comment_range(sheet, time)
+    write_range = get_night_comment_range(sheet, time, last_night_time)
     if not write_range:
         return
     logging.info(f'Comment for the last night wake-up. Log into range {write_range}')
-    last_night_time = get_last_night_time(sheet, time)
+    # last_night_time = get_last_night_time(sheet, time)
     sheet.update(write_range, [[cmt]])
     await message.reply(f'Добавили комментарий к просыпанию в {last_night_time}', reply=False)
 
@@ -109,13 +142,18 @@ async def echo(message: types.Message):
     if match:
         sid = match.groups(1)[0]
         user_id = message.from_user.id
-        with open('user_sid.json', 'r+') as f:
-            data = json.load(f)
-            if user_id in data:
-                data.pop(user_id)
-            data[user_id] = sid
-            f.seek(0)
-            json.dump(data, f, indent=4)
+        if not os.path.exists('user_sid.json'):
+            with open('user_sid.json', 'w') as fw:
+                data = {user_id: sid}
+                json.dump(data, fw, indent=4)
+        else:
+            with open('user_sid.json', 'r+') as f:
+                data = json.load(f)
+                if user_id in data:
+                    data.pop(user_id)
+                data[user_id] = sid
+                f.seek(0)
+                json.dump(data, f, indent=4)
         await message.reply('Идентификатор таблицы сохранен', reply=False)
     else:
         await message.reply(f'Для записи комментария к НП перейдите в режим /night и добавьте НП.')
@@ -140,14 +178,17 @@ def get_night_wake_range(sheet: Sheet, time: datetime):
     return write_range
 
 
-def get_night_comment_range(sheet: Sheet, time: datetime):
+def get_night_comment_range(sheet: Sheet, time: datetime, time_to_edit: str):
     dates = sheet.get_data(sheet_range=DATES_RANGE)
     col_to_write = dates[0].index(time.strftime('%d.%m.%Y'))
     filled_col = sheet.get_data(sheet_range=_get_night_range(RANGE, col_to_write))
-    last_row_ix = len(filled_col)
-    if last_row_ix == 45:
+    fc = filled_col[NIGHT_ROW_INDEX:]
+    logging.info(f'{time_to_edit=}, {fc=}')
+    row_ix = len(filled_col) if not time_to_edit else filled_col.index([time_to_edit]) + 1
+    # last_row_ix = len(filled_col)
+    if row_ix == 45:
         return None
-    write_range = _get_night_write_range(RANGE, col_to_write + 1, last_row_ix - 1)
+    write_range = _get_night_write_range(RANGE, col_to_write + 1, row_ix - 1)
     return write_range
